Support iminuit 2 only
