+++ gammapy/modeling/fit.py
@@ -5,7 +5,7 @@ import numpy as np
 from gammapy.utils.table import table_from_row_data
 from gammapy.utils.pbar import progress_bar
 from .covariance import Covariance
-from .iminuit2 import confidence_iminuit, covariance_iminuit, mncontour, optimize_iminuit
+from .iminuit import confidence_iminuit, covariance_iminuit, mncontour, optimize_iminuit
 from .scipy import confidence_scipy, optimize_scipy
 from .sherpa import optimize_sherpa
